+++ builders/RedisBuilder.d.ts
@@ -2,16 +2,16 @@ import type { RedisConfig, RedisPipeline, RedisSubscribe } from "../types/redis"
 export default class RedisBuilder {
     private static clients;
     private static emitter;
-    static setClient(cfg: RedisConfig): Record<string, Function>;
+    static setClient(cfg: RedisConfig, name?: string): Record<string, Function>;
     static connection(name: string): Record<string, Function>;
     static connect(name?: string): Promise<Bun.RedisClient>;
     static disconnect(name?: string): Promise<void>;
-    static get(key: Bun.RedisClient.KeyLike, connection?: string): Promise<any>;
-    static set(key: Bun.RedisClient.KeyLike, value: any, ttl?: number, connection?: string): Promise<number | "OK">;
-    static del(key: Bun.RedisClient.KeyLike, connection?: string): Promise<number>;
+    static get(key: Bun.RedisClient.KeyLike, connection?: string, disconnectAfter?: boolean): Promise<any>;
+    static set(key: Bun.RedisClient.KeyLike, value: any, ttl?: number, connection?: string, disconnectAfter?: boolean): Promise<number | "OK">;
+    static del(key: Bun.RedisClient.KeyLike, connection?: string, disconnectAfter?: boolean): Promise<number>;
     static publish(channel: string, message: any, connection?: string): Promise<number>;
     static subscribe(channel: string, listener: Bun.RedisClient.StringPubSubListener, connection?: string): Promise<RedisSubscribe>;
-    static pipeline(fn: (pipe: RedisPipeline) => void, connection?: string): Promise<any[]>;
+    static pipeline(fn: (pipe: RedisPipeline) => void, connection?: string, disconnectAfter?: boolean): Promise<any[]>;
     static on(event: "connect" | "disconnect" | "error", listener: (...args: Array<any>) => void): void;
     static off(event: "connect" | "disconnect" | "error", listener: (...args: Array<any>) => void): void;
     private static get config();